Part 2 completed
I changed a bit the package diagram, I removed the "search page" as we
don't really have a search page, it's just a box that prompts the search
manager. The page taht shows the calendar of other people is not so
different from the "calendar page" so I have united them.